+++ middleware/pengesahanToken.js
@@ -4,15 +4,15 @@ const config = process.env;
 
 const pengesahanToken = (req, res, next) => {
     // Dapatkan header request
-    const headers = req.headers['authorization'];
-    const bearer = headers.split(' ');
+    const authorization = req.headers['authorization'];
+    const bearer = authorization.split(' ');
         
     // Dapatkan token
     const token = bearer[1];
 
     // Memastikan token wujud
     if(!token) {
-        return res.status(403).send({ mesej: 'Token diperlukan'});
+        return res.status(400).send({ mesej: 'Token diperlukan'});
     }
 
     try {
@@ -21,11 +21,10 @@ const pengesahanToken = (req, res, next) => {
 
         // Mengumpukkan nilai token
         req.pengguna = nyahsulit;
-        console.log(req.pengguna)
 
     } catch (err) {
         // Token tidak sah
-        return res.status(401).send({ mesej: 'Token tidak sah'});
+        return res.status(403).send({ mesej: 'Token tidak sah'});
     }
     return next();
 }
